56_name_card/requirements.txt
Flask==3.0.3
flask-babel==4.0.0
python-dotenv==1.0.1